Refactoring of Http dispatcher
