+++ dotnet/src/dotnetcore/GxClasses.Web/GxClasses.Web.csproj
@@ -6,7 +6,7 @@
 		<PackageId>GeneXus.Classes.Web.Core</PackageId>
 		<PackageTags>GXGeolocation GXHttpServices WebSocket</PackageTags>
 		<PackageId>GeneXus.Classes.Web.Core</PackageId>
-		<NoWarn>CA1812</NoWarn>
+		<NoWarn>CA1812,CA1819</NoWarn>
 	</PropertyGroup>
 
   <ItemGroup>
